task and comment stuff

+++ Data/Models/Comment.cs
@@ -9,8 +9,8 @@ namespace pm.Models
     {
         public long Id { get; set; }
         public string Content { get; set; }
-        public int UserId { get; set; }
-        public int TaskId { get; set; }
+        public long UserId { get; set; }
+        public long TaskId { get; set; }
         public DateTime CreationDate { get; set; }
     }
 }
